bot/requirements.txt
python-telegram-bot[aio]==20.6
firebase-admin==6.0.1
python-dotenv==1.0.0
pydantic==1.10.7
aiohttp==3.9.5
uvicorn==0.22.0
fastapi==0.95.2
redis==4.7.0
async-timeout==4.0.2
python-dateutil==2.8.2
pytest==7.4.0
httpx==0.24.1
